+++ VueAPI/src/main/java/com/ssafy/vue/service/HouseMapService.java
@@ -12,9 +12,9 @@ public interface HouseMapService {
 	List<SidoGugunCodeDto> getGugunInSido(String sido) throws Exception;
 	List<HouseInfoDto> getDongInGugun(String gugun) throws Exception;
 	List<HouseInfoDto> getAptInDong(String dong) throws Exception;
-	boolean countUpHouse(int aptcode) throws Exception;
+	boolean countUpHouse(long aptcode) throws Exception;
 	List<HouseInfoDto> popularHouse() throws Exception;
-	List<HouseDealDto> getAptDeal(int aptCode);
+	List<HouseDealDto> getAptDeal(long aptCode);
 	List<HouseInfoDto> getNewDeal() throws Exception;
 	
 }
